tracking-service/requirements.txt
Django>=4.0
restframework
corsheaders
requests
kafka-python

